Reorg server Tweak config files Consolidate all model files in models directory
