+++ server/package.json
@@ -8,7 +8,7 @@
     "build": "tsc",
     "start": "node dist/server.js",
     "watch": "nodemon dist/server.js",
-    "seed": "node dist/seeds/seed.js"
+    "seed": "node dist/seeds/index.js"
   },
   "author": "",
   "license": "ISC",
@@ -26,6 +26,7 @@
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
+    "@types/express": "^4.17.22",
     "@types/jsonwebtoken": "^9.0.9",
     "@types/node": "^22.15.29",
     "nodemon": "^3.1.4",
