--- server/src/models/index.js
@@ -1,4 +1,0 @@
-// import Tech from './Tech.js';
-// import Matchup from './Matchup.js';
-export {};
-// export default { Tech, Matchup };
