+++ server/src/models/photo/photoSchema.ts
@@ -1,6 +1,6 @@
 // schemas/Photo.ts
 import mongoose from 'mongoose';
-import {IPhoto} from './typephoto';
+import { IPhoto } from './photoType.js';
 
 const PhotoSchema = new mongoose.Schema({
   title: { type: String, required: true },
