+++ server/tsconfig.json
@@ -15,6 +15,5 @@
       "rootDir": "src",
       "outDir": "dist"
     },
-    "include": ["src"]
-  
+    "include": ["src"] 
 }
