+++ tsconfig.json
@@ -26,5 +26,6 @@
     "noImplicitReturns": true /* Enable error reporting for codepaths that do not explicitly return in a function. */,
     "noImplicitOverride": true /* Ensure overriding members in derived classes are marked with an override modifier. */,
     "useUnknownInCatchVariables": true /* Ensure 'unknown' is used for catch clause variables. */
-  }
+  },
+  "exclude": ["server/dist", "client/dist"]
 }
